Limit outfit pieces to 4 in each outfit preview
