+++ src/components/OutfitDetails/OutfitDetails.js
@@ -1,9 +1,23 @@
 import React, { useEffect } from 'react';
 import { useParams } from 'react-router-dom';
 import ErrorMessage from "../ErrorMessage/ErrorMessage";
+import { getOutfitPieces } from '../../apiCalls';
 
 const OutfitDetails = ({ user, setAppError, appError }) => {
   const outfitID = useParams().id;
+
+  useEffect(() => {
+    const apiCall = async (userID, outfitID) => {
+      try {
+        const data = await getOutfitPieces(userID, outfitID)
+        console.log(data)
+      } catch (error) {
+        //handle errpr
+      }
+      
+    }
+    // apiCall(user.userID, outfitID)
+  },[])
 
   return (
     <>
